requirements.txt
alembic==1.16.2
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
DateTime==5.5
Flask==3.1.1
Flask-CKEditor==1.0.0
Flask-WTF==1.2.2
google-api-core==2.28.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-core==2.5.0
google-cloud-storage==3.7.0
google-crc32c==1.7.1
google-resumable-media==2.8.0
googleapis-common-protos==1.70.0
gspread==6.2.1
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
pillow_heif==1.1.0
proto-plus==1.26.1
protobuf==6.31.1
psycopg==3.3.2
psycopg-binary==3.3.2
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyparsing==3.2.3
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==80.9.0
soupsieve==2.7
SQLAlchemy==2.0.41
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
webencodings==0.5.1
Werkzeug==3.1.3
WTForms==3.2.1
zope.interface==7.2
